+++ src/main/java/moviestreamingservice/domain/hall/HallService.java
@@ -3,6 +3,9 @@ package moviestreamingservice.domain.hall;
 import lombok.RequiredArgsConstructor;
 import moviestreamingservice.domain.cinema.Cinema;
 import moviestreamingservice.domain.cinema.CinemaRepository;
+import moviestreamingservice.domain.hall.dto.HallRequest;
+import moviestreamingservice.domain.hall.dto.HallResponse;
+import moviestreamingservice.exception.NotFoundException;
 import moviestreamingservice.utilities.SeatUtil;
 import org.springframework.stereotype.Service;
 
@@ -13,26 +16,38 @@ import java.util.List;
 public class HallService {
     private final HallRepository hallRepository;
     private final CinemaRepository cinemaRepository;
-    public List<Hall> getHallsByCinema(Long cinemaId) {
-        Cinema cinema = cinemaRepository.findById(cinemaId).orElseThrow(() -> new RuntimeException("Cinema not found"));
-        return cinema.getHalls();
+    public List<HallResponse> getHallsByCinema(Long cinemaId) {
+        Cinema cinema = cinemaRepository.findById(cinemaId).orElseThrow(() -> new NotFoundException("Cinema not found"));
+        return cinema.getHalls().stream()
+                .map(HallMapper::mapHallToResponse).toList();
     }
-    public Hall getHall(Long id) {
-        return hallRepository.findById(id).orElseThrow(()->new RuntimeException("Hall not found"));
+    public HallResponse getHall(Long id) {
+        Hall hall = hallRepository.findById(id).orElseThrow(()->new NotFoundException("Hall not found"));
+        return HallMapper.mapHallToResponse(hall);
     }
-    public Hall createHall(Long cinemaId, Hall hall) {
-        Cinema cinema = cinemaRepository.findById(cinemaId).orElseThrow(()->new RuntimeException("Cinema not found"));
+    public HallResponse createHall(Long cinemaId, HallRequest hallRequest) {
+        Cinema cinema = cinemaRepository.findById(cinemaId).orElseThrow(()->new NotFoundException("Cinema not found"));
+        Hall hall = new Hall();
+
+        hall.setHallNumber(hallRequest.hallNumber());
+        hall.setTotalSeats(hallRequest.totalSeats());
         hall.setCinema(cinema);
-        hall.setSeats(SeatUtil.generateSeats(hall.getTotalSeats()));
-        return hallRepository.save(hall);
+        hall.setSeats(SeatUtil.generateSeats(hallRequest.totalSeats()));
+
+        Hall saved = hallRepository.save(hall);
+        return HallMapper.mapHallToResponse(saved);
     }
-    public Hall updateHall(Long id, Hall updated) {
-        Hall hall = getHall(id);
-        hall.setHallNumber(updated.getHallNumber());
-        hall.setTotalSeats(updated.getTotalSeats());
-        return hallRepository.save(hall);
+    public HallResponse updateHall(Long id, HallRequest hallRequest) {
+        Hall hall = hallRepository.findById(id).orElseThrow(()-> new NotFoundException("Hall not found"));
+        hall.setHallNumber(hallRequest.hallNumber());
+        hall.setTotalSeats(hallRequest.totalSeats());
+        hall.setSeats(SeatUtil.generateSeats(hallRequest.totalSeats()));
+        return  HallMapper.mapHallToResponse(hallRepository.save(hall));
     }
     public void deleteHall(Long id) {
+        if(!hallRepository.existsById(id)){
+            throw new NotFoundException("Hall not found");
+        }
         hallRepository.deleteById(id);
     }
 }
