+++ src/main/java/moviestreamingservice/domain/movie/MovieService.java
@@ -1,6 +1,10 @@
 package moviestreamingservice.domain.movie;
 
 import lombok.RequiredArgsConstructor;
+import moviestreamingservice.domain.movie.dto.MovieRequest;
+import moviestreamingservice.domain.movie.dto.MovieResponse;
+import moviestreamingservice.exception.NotFoundException;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -9,26 +13,48 @@ import java.util.List;
 @RequiredArgsConstructor
 public class MovieService {
     private final MovieRepository movieRepository;
-    public Movie createMovie(Movie movie){
-        return movieRepository.save(movie);
+    public MovieResponse createMovie(MovieRequest movieRequest) {
+        Movie movie = MovieMapper.toEntity(movieRequest);
+        return MovieMapper.toResponse(movieRepository.save(movie));
     }
-    public Movie updateMovie(Long id, Movie updated){
+    public MovieResponse updateMovie(Long id, MovieRequest updatedReq){
         Movie movie = movieRepository.findById(id)
-                .orElseThrow(()-> new RuntimeException("Movie not found"));
-        movie.setTitle(updated.getTitle());
-        movie.setOverview(updated.getOverview());
-        movie.setPosterUrl(updated.getPosterUrl());
-        movie.setGenres(updated.getGenres());
-        return movieRepository.save(movie);
+                .orElseThrow(() -> new NotFoundException("Movie not found"));
+
+        movie.setTitle(updatedReq.title());
+        movie.setOverview(updatedReq.overview());
+        movie.setPosterUrl(updatedReq.posterUrl());
+        movie.setBackdropUrl(updatedReq.backdropUrl());
+        movie.setVoteAverage(updatedReq.voteAverage());
+        movie.setVoteCount(updatedReq.voteCount());
+        movie.setPopularity(updatedReq.popularity());
+        movie.setReleaseDate(updatedReq.releaseDate());
+        movie.setGenres(updatedReq.genres());
+
+        Movie saved = movieRepository.save(movie);
+
+        return MovieMapper.toResponse(saved);
     }
     public void deleteMovie(Long id) {
+        if(!movieRepository.existsById(id)) {
+            throw new NotFoundException("Movie not found");
+        }
         movieRepository.deleteById(id);
     }
-    public List<Movie> getAll() {
-        return movieRepository.findAll();
+    public List<MovieResponse> getAll() {
+        return movieRepository.findAll().stream()
+                .map(MovieMapper::toResponse)
+                .toList();
     }
-    public Movie getById(Long id) {
-        return movieRepository.findById(id)
-                .orElseThrow(()-> new RuntimeException("Movie not found"));
+    public MovieResponse getById(Long id) {
+        Movie movie = movieRepository.findById(id)
+                .orElseThrow(()-> new NotFoundException("Movie not found"));
+        return MovieMapper.toResponse(movie);
+    }
+    public List<MovieResponse> getAll(Pageable pageable) {
+        return movieRepository.findAll(pageable)
+                .stream()
+                .map(MovieMapper::toResponse)
+                .toList();
     }
 }
